+++ Modules/AdvancedSecurity/Private/Block-FingerProtocol.ps1
@@ -21,7 +21,7 @@ function Block-FingerProtocol {
         
     .NOTES
         Author: NexusOne23
-        Version: 2.2.1
+        Version: 2.2.2
         Requires: Administrator privileges
         
         REFERENCES: