+++ Modules/AdvancedSecurity/Private/Set-SRPRules.ps1
@@ -27,7 +27,7 @@ function Set-SRPRules {
         
     .NOTES
         Author: NexusOne23
-        Version: 2.2.1
+        Version: 2.2.2
         Requires: Administrator privileges
         
         REFERENCES: