+++ Modules/EdgeHardening/EdgeHardening.psm1
@@ -16,7 +16,7 @@
     
 .NOTES
     Author: NexusOne23
-    Version: 2.2.1
+    Version: 2.2.2
     Requires: PowerShell 5.1+, Administrator privileges
 #>