+++ Modules/Privacy/Privacy.psm1
@@ -16,7 +16,7 @@
 
 .NOTES
     Module: Privacy
-    Version: 2.2.1
+    Version: 2.2.2
     Author: NoID Privacy
 #>